pyproject.toml
[tool.poetry]
name = "airhorn-supremacy"
version = "0.1.0"
description = ""
authors = ["TAG-Epic <tagepicuwu@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
botconfig = "^1.0.1"
"discord.py" = {extras = ["voice"], version = "^1.6.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
